removed base classes since we don't need them without async version
